\name{surv2sample-internal}
\alias{mini.survival}
\alias{mini.agreg.fit}
\alias{mini.coxph.detail}
\alias{mini.survfit.coxph}
\alias{allsubsets}
\alias{globnestedsubsets}
\alias{print.smoothproptest}
\alias{print.scoreproptest}
\title{Internal Functions for the Package proptest}
\description{
Internal functions for the package \pkg{proptest}.
Usually not to be directly called by users.
}
\author{David Kraus (\url{http://www.davidkraus.net/})}
\keyword{internal}
